worth-ui 3.16.1 phase 1: reach the scrolling endpoint
Implement the smallest complete owner-to-presentation scrolling path and
run it on one real Pulse panel: both axes, moving thumbs, coarse-wheel
smoothing through the Motion sampled lane, direct thumb dragging, and a
working click after movement, with pixel and geometry assertions from the
first step.

Runtime
- Scroll succession is prepared before publication and committed only
  after the geometry that produces the pixels is accepted. Both wheel
  branches and chrome placement apply poses first and install the
  successor on success; a refusal is a typed Geometry denial that leaves
  the owner, the displayed pose and the thumb untouched.
- A smooth notch stages its transition target on the successor clone and
  publishes it through Motion; the clone is installed only when the
  publication succeeds, so an unpublished settle leaves no orphaned target
  and a second notch inside the horizon retargets from the first target.
- Accepted Motion samples settle the semantic offset through a typed
  UiScrollSettleDisposition (Applied, DeferredPresentationInFlight,
  Refused{Geometry, Resolution, WriteBack}, Idle); a deferred settle is
  owed and paid by the next frame with the same accepted translation.
  The accepted offset is the content box at rest minus the sampled
  position, proven at the Pulse content box rather than the origin.
- Ticks that sample only scrolled content owe the host no sample work;
  the displacement is geometry the settle re-lowers.
- Declared line extents, smooth wheel behaviour, chrome contracts and
  host wheel-line normalization lower through the dsl, host-contract,
  host-native and headless crates; smooth policy without Motion is
  refused at normalization.

Pulse
- Dashboard scroll region, chrome declarations and native journey
  (native_wheel_notch_and_thumb_drag) with scroll-chrome pixel
  adjudication and progression evidence.

Evidence
- Integrated appearance World gained a Scroll-installed fixture with
  facade-level tests for commit-after-evidence, staging, frame settle,
  deferral and retargeting; runtime lib 1799 passed / 0 failed /
  2 ignored; clippy -D warnings clean; fmt, line caps, boundary and
  agent-context checks pass. Both independent reviewer gates cleared.

+++ workspaces/worth-ui/crates/worth-ui-runtime/src/facade/entry/active_application_session/mounted_scroll_geometry_tests.rs
@@ -43,7 +43,7 @@ fn scroll_pose_is_local_idempotent_and_rejects_a_later_invalid_owner_before_effe
     ));
     let offset = crate::runtime::scroll::UiScrollOffset::new(
         0,
-        8 * worth_ui_host_contract::UI_HOST_SURFACE_POSITION_SUBPIXELS_PER_UNIT as i64,
+        8 * worth_ui_host_contract::UI_HOST_SURFACE_POSITION_SUBPIXELS_PER_UNIT,
     )
     .unwrap();
     assert!(!session.mounted.projection_changes_pending());
